Reopen solved post (#57)
* Choosing Best Answer

Author is able to choose best answer from comments, which automatically places the post as solved and removes from the currently open post list. Closes #40

* Travis Test Mod

Modified Travis to account for new variable "SOLVED" under the Model "POSTS"

* Reopen posts marked solved

Added a button and method to reopen posts that are marked as solved by the author only. Related to #40