25 google facebook api (#59)
* Facebook Login

Implemented the Facebook register/login feature. To be tested by others.

* Google and Facebook Login Implementation

Implementation of both logins working dynamically. Meaning that they use the same functions with the difference in providers only. This means that adding log in with github and/or twitter will be very easily if we want to in the future.

* Final code cleaning

Final code cleaning, and slight login page adjustments.
Closes #25